requirements.txt
python-telegram-bot==20.7
PyYAML==6.0.1
requests==2.31.0